requirements.txt
telethon
streamlit
pandas
langdetect
tqdm
backoff
python-dotenv